apps/api/pyproject.toml
[project]
name = "tibetan-urine-ai-api"
version = "0.1.0"
description = "API service for TibetanUrineAI"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "Pillow>=10.0.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic-settings>=2.4.0",
  "python-multipart>=0.0.9"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.27.0"
]

[build-system]
requires = ["setuptools>=70.0.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
